feat: fold MAC/hostname into health endpoint, remove WoL service
- Add machineName + macAddresses to HealthResponse
- Inline MAC enumeration in HealthEndpoints
- Delete WolService and WolServiceTests
- Remove /system/mac and /system/wol endpoints
- Keep ConfigPaths persistence fix (ProgramData)
