+++ src/HaPcRemote.Service/AppJsonContext.cs
@@ -23,7 +23,6 @@ namespace HaPcRemote.Service;
 [JsonSerializable(typeof(ApiResponse<List<MonitorInfo>>))]
 [JsonSerializable(typeof(MacAddressInfo))]
 [JsonSerializable(typeof(List<MacAddressInfo>))]
-[JsonSerializable(typeof(ApiResponse<List<MacAddressInfo>>))]
 [JsonSourceGenerationOptions(
     PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase,
     DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull)]
