+++ src/HaPcRemote.Service/Endpoints/SystemEndpoints.cs
@@ -25,36 +25,6 @@ public static class SystemEndpoints
             }
         });
 
-        group.MapGet("/mac", () =>
-        {
-            var macs = WolService.GetMacAddresses();
-            return Results.Json(ApiResponse.Ok(macs),
-                AppJsonContext.Default.ApiResponseListMacAddressInfo);
-        });
-
-        group.MapPost("/wol/{mac}", async (string mac, ILogger<IPowerService> logger) =>
-        {
-            try
-            {
-                logger.LogInformation("WoL requested for MAC: {Mac}", mac);
-                await WolService.SendWolAsync(mac);
-                return Results.Json(ApiResponse.Ok($"WoL packet sent to {mac}"),
-                    AppJsonContext.Default.ApiResponse);
-            }
-            catch (ArgumentException ex)
-            {
-                logger.LogWarning(ex, "Invalid MAC address: {Mac}", mac);
-                return Results.Json(ApiResponse.Fail(ex.Message), AppJsonContext.Default.ApiResponse,
-                    statusCode: StatusCodes.Status400BadRequest);
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex, "Failed to send WoL packet to {Mac}", mac);
-                return Results.Json(ApiResponse.Fail(ex.Message), AppJsonContext.Default.ApiResponse,
-                    statusCode: StatusCodes.Status500InternalServerError);
-            }
-        });
-
         return group;
     }
 }
